divide verification action
